% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scrape}
\alias{scrape}
\title{Scrape data from web pages}
\usage{
scrape(urls, url_file = NULL, ratelimit = 3, scraper = "generic_open",
  args = list(), outdir = NULL)
}
\arguments{
\item{urls}{A vector of URLs to scrape}

\item{scraper}{A single scraper to use, or list of scrapers the same length
as \code{urls}.  If \code{NULL}, \code{scrape} will choose scrapers based on
the domains of URLs}

\item{list}{The form to return results in, either "list", "data.frame", or
"none" to only retain results as JSON files on disk}

\item{outdir}{The directory to write results to.}
}
\description{
Take a vector of URLs and scrape data from the associated web pages, using
\code{quickscrape}
}

